fix : transaction 어노테이션 제거 및 transaction 처리를 위한 서비스 클래스 분리, synchronized 위치 변경
